WOM-281: Rename module to follow convention and other changes
- Module name must match Gitlab project slug to appease Marvin.
- Updated poetry dependencies for various make targets to work in
devcontainer.
- Added sphinx docs with correct structure.
- Linting fixes, but only on tests as first pass.

+++ use_poetry.sh
@@ -1,4 +1,4 @@
-[[ -s pyproject.toml ]] && {
+if [ -s pyproject.toml ]; then
     rm pyproject.toml;
     ln -s pyproject_for_poetry.toml pyproject.toml; 
-}
+fi
