+++ use_setup.sh
@@ -1,4 +1,4 @@
-[[ -s pyproject.toml ]] && {
+if [ -s pyproject.toml ]; then
     rm pyproject.toml;
     ln -s pyproject_for_setup.toml pyproject.toml; 
-}
+fi
